+++ src/SFA.DAS.LevyTransferMatching.Web/Models/Opportunities/Opportunity.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.LevyTransferMatching.Web.Models.SearchFunding
+﻿namespace SFA.DAS.LevyTransferMatching.Web.Models.Opportunities
 {
     public class Opportunity
     {
