+++ src/SFA.DAS.LevyTransferMatching.Web/Orchestrators/IOpportunitiesOrchestrator.cs
@@ -1,5 +1,5 @@
-﻿using SFA.DAS.LevyTransferMatching.Web.Models.SearchFunding;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
+using SFA.DAS.LevyTransferMatching.Web.Models.Opportunities;
 
 namespace SFA.DAS.LevyTransferMatching.Web.Orchestrators
 {
