+++ src/SFA.DAS.LevyTransferMatching.Web/Views/Opportunities/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model SFA.DAS.LevyTransferMatching.Web.Models.SearchFunding.IndexViewModel
+﻿@model SFA.DAS.LevyTransferMatching.Web.Models.Opportunities.IndexViewModel
 
 @{
     ViewData["Title"] = "Search funding opportunities";
